+++ CHANGES.rst
@@ -1,3 +1,4 @@
+* Changed MultiValueField to run any validators it was given.
 * Changed URL and email address validators to handle IDNA domains.
 * Changed ``CheckboxInput`` to correctly handle ``0`` as a value.
 * Added ``BaseFormSet.prototype.hasChanged()``